+++ icmplib/multiping.py
@@ -86,7 +86,7 @@ async def async_multiping(addresses, count=2, interval=0.5, timeout=2,
     :type interface: str, optional
     :param interface: The network interface to bind to (e.g., 'eth0',
         'wlan0'). By default, the socket is not bound to a specific
-        interface. Only available on Unix systems. Ignored on Windows.
+        interface. Only available on Linux. Ignored on macOS and Windows.
 
     Advanced (**kwags):
 
@@ -222,7 +222,7 @@ def multiping(addresses, count=2, interval=0.5, timeout=2,
     :type interface: str, optional
     :param interface: The network interface to bind to (e.g., 'eth0',
         'wlan0'). By default, the socket is not bound to a specific
-        interface. Only available on Unix systems. Ignored on Windows.
+        interface. Only available on Linux. Ignored on macOS and Windows.
 
 
     Advanced (**kwags):