+++ icmplib/ping.py
@@ -85,7 +85,7 @@ def ping(address, count=4, interval=1, timeout=2, id=None, source=None,
     :type interface: str, optional
     :param interface: The network interface to bind to (e.g., 'eth0',
         'wlan0'). By default, the socket is not bound to a specific
-        interface. Only available on Unix systems. Ignored on Windows.
+        interface. Only available on Linux. Ignored on macOS and Windows.
 
     Advanced (**kwags):
 
@@ -222,7 +222,7 @@ async def async_ping(address, count=4, interval=1, timeout=2, id=None,
     :type interface: str, optional
     :param interface: The network interface to bind to (e.g., 'eth0',
         'wlan0'). By default, the socket is not bound to a specific
-        interface. Only available on Unix systems. Ignored on Windows.
+        interface. Only available on Linux. Ignored on macOS and Windows.
 
 
     Advanced (**kwags):